Fix comp warning
git-svn-id: http://svn.cern.ch/guest/lhcb/Online/trunk@159184 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: 3cc2dd73825ac2328ed6b548046cb5b8957b5649

+++ Online/Online/OnlineBase/src/MBM/Summary.cpp
@@ -78,8 +78,9 @@ int Summary::show(BufferMemory* dsc, bool show_states)   {
   ::sprintf(txt," Buffer \"%s\"",dsc->bm_name);
   ::printf("%-26s  Events: Produced:%ld Seen:%ld Pending:%ld Max:%d\n",
 	   txt, long(ctr->tot_produced), long(ctr->tot_seen), ctr->i_events, ctr->p_emax);
-  ::printf("%-26s  Space(kB):[Tot:%d Free:%ld] Users:[Tot:%ld Max:%d]\n\n",
-	   "",(ctr->bm_size*ctr->bytes_p_Bit)/1024, (ctr->i_space*ctr->bytes_p_Bit)/1024, 
+  ::printf("%-26s  Space(kB):[Tot:%ld Free:%ld] Users:[Tot:%ld Max:%ld]\n\n",
+	   "",long((ctr->bm_size*ctr->bytes_p_Bit)/1024), 
+	   long((ctr->i_space*ctr->bytes_p_Bit)/1024), 
 	   ctr->i_users, ctr->p_umax);
 
   bool first = true;
